refactor(react-components): SetPlanRow を分割・命名整理し、edit/weightStep を必須化
レビュー指摘:
- weightStep / repsStep のオプショナル理由が薄い（種目スキーマから受け取る前提）
- editable じゃないパターンは想定外（edit を optional にする理由なし）
- drawer / popover を別ファイルに切り出す
- params という型・変数名は意味が薄い（→ 案B: SetPlanPattern + 判別子は kind）
- useEditingState / EditFormContent も別ファイル
- repsStep は「常に 1 で十分」が確認できたため API から削除しハードコード

対応:
- program-detail/set-plan-row/ にフォルダ化:
  - index.tsx (SetPlanRow + formatPattern)
  - set-plan-row-popover.tsx (SetPlanRowPopover)
  - set-plan-row-drawer.tsx (SetPlanRowDrawer)
  - set-plan-edit-form.tsx (SetPlanEditForm + DraftFields)
  - use-set-plan-editing.ts (useSetPlanEditing + SetPlanPattern + Draft + helpers)
- 型 Params → SetPlanPattern、判別子 pattern → kind（型と field の衝突回避）
- 変数 / field params → pattern
- weightStep を必須プロップに（exercises スキーマの値を caller が渡す）
- repsStep は API から削除し step={1} ハードコード
- edit: { title; onChange } を必須プロップに
- ExercisePlanSection の Exercise type に weightStep を追加し、ProgramDetail で exercise から取り出す（プレースホルダ削除）
- exercise === null のとき SetPlanSection を出さない
- ProgramDetail に onSetPlanChange プロップを追加（SetPlanRow 編集の伝搬）
- PATTERN_OPTIONS と patternLabel の重複定義を整理（PATTERN_KINDS + patternLabel に分離）
