+++ packages/react-components/src/views/program-detail/index.stories.tsx
@@ -11,24 +11,28 @@ const benchPress: ExercisePlan["exercise"] = {
 	id: "ex-bench",
 	name: "ベンチプレス",
 	weightUnit: "kg",
+	weightStep: 2.5,
 };
 
 const inclineDumbbell: ExercisePlan["exercise"] = {
 	id: "ex-incline-db",
 	name: "インクラインダンベルプレス",
 	weightUnit: "kg",
+	weightStep: 1,
 };
 
 const tricepsPushdown: ExercisePlan["exercise"] = {
 	id: "ex-pushdown",
 	name: "トライセプスプッシュダウン",
 	weightUnit: "kg",
+	weightStep: 2.5,
 };
 
 const squat: ExercisePlan["exercise"] = {
 	id: "ex-squat",
 	name: "バックスクワット",
 	weightUnit: "kg",
+	weightStep: 2.5,
 };
 
 const SAMPLE_DAYS: Day[] = [
@@ -43,15 +47,15 @@ const SAMPLE_DAYS: Day[] = [
 				setPlans: [
 					{
 						id: "sp-d1-bench-1",
-						params: { pattern: "weight-x-reps", weight: 100, reps: 5 },
+						pattern: { kind: "weight-x-reps", weight: 100, reps: 5 },
 					},
 					{
 						id: "sp-d1-bench-2",
-						params: { pattern: "weight-x-reps", weight: 100, reps: 5 },
+						pattern: { kind: "weight-x-reps", weight: 100, reps: 5 },
 					},
 					{
 						id: "sp-d1-bench-3",
-						params: { pattern: "weight-x-rpe", weight: 100, rpe: 9 },
+						pattern: { kind: "weight-x-rpe", weight: 100, rpe: 9 },
 					},
 				],
 			},
@@ -61,11 +65,11 @@ const SAMPLE_DAYS: Day[] = [
 				setPlans: [
 					{
 						id: "sp-d1-incline-1",
-						params: { pattern: "weight-x-reps", weight: 30, reps: 10 },
+						pattern: { kind: "weight-x-reps", weight: 30, reps: 10 },
 					},
 					{
 						id: "sp-d1-incline-2",
-						params: { pattern: "weight-x-reps", weight: 30, reps: 10 },
+						pattern: { kind: "weight-x-reps", weight: 30, reps: 10 },
 					},
 				],
 			},
@@ -75,11 +79,11 @@ const SAMPLE_DAYS: Day[] = [
 				setPlans: [
 					{
 						id: "sp-d1-pushdown-1",
-						params: { pattern: "reps-x-rpe", reps: 12, rpe: 8 },
+						pattern: { kind: "reps-x-rpe", reps: 12, rpe: 8 },
 					},
 					{
 						id: "sp-d1-pushdown-2",
-						params: { pattern: "reps-x-rpe", reps: 12, rpe: 8 },
+						pattern: { kind: "reps-x-rpe", reps: 12, rpe: 8 },
 					},
 				],
 			},
@@ -96,15 +100,15 @@ const SAMPLE_DAYS: Day[] = [
 				setPlans: [
 					{
 						id: "sp-d2-squat-1",
-						params: { pattern: "weight-x-reps", weight: 140, reps: 3 },
+						pattern: { kind: "weight-x-reps", weight: 140, reps: 3 },
 					},
 					{
 						id: "sp-d2-squat-2",
-						params: { pattern: "weight-x-reps", weight: 140, reps: 3 },
+						pattern: { kind: "weight-x-reps", weight: 140, reps: 3 },
 					},
 					{
 						id: "sp-d2-squat-3",
-						params: { pattern: "weight-x-reps", weight: 140, reps: 3 },
+						pattern: { kind: "weight-x-reps", weight: 140, reps: 3 },
 					},
 				],
 			},
@@ -134,6 +138,7 @@ const meta = {
 	args: {
 		defaultSelectedDayId: "d1",
 		onAddDay: fn(),
+		onSetPlanChange: fn(),
 	},
 	decorators: [
 		(Story) => (
@@ -171,7 +176,7 @@ export const LongProgramName: Story = {
 	},
 };
 
-// 設計判断 #58 によりアプリ層が初期構造（Day×1 + 種目計画×1（種目未確定） + セット計画×1（params: null））を渡す。
+// 設計判断 #58 によりアプリ層が初期構造（Day×1 + 種目計画×1（種目未確定） + セット計画×1（pattern: null））を渡す。
 // プログラム新規作成 → Day新規作成 → プログラム詳細 のフロー直後を表示する想定の状態。
 // 種目選択 picker は本ビューには持たず、種目選択 UI 自体はプログラム新規作成・Day 新規作成・種目計画追加タスクで導入する想定。
 export const InitialStateAfterCreation: Story = {
@@ -187,7 +192,7 @@ export const InitialStateAfterCreation: Story = {
 					{
 						id: "ep-init",
 						exercise: null,
-						setPlans: [{ id: "sp-init", params: null }],
+						setPlans: [{ id: "sp-init", pattern: null }],
 					},
 				],
 			},
