+++ packages/react-components/src/views/program-detail/set-plan-section.tsx
@@ -1,14 +1,23 @@
 import type { ComponentProps, FC } from "react";
 import { SetPlanRow } from "./set-plan-row";
 
-type Params = ComponentProps<typeof SetPlanRow>["params"];
+type Pattern = ComponentProps<typeof SetPlanRow>["pattern"];
 
 type Props = {
-	setPlans: { id: string; params: Params }[];
+	setPlans: { id: string; pattern: Pattern }[];
 	weightUnit: "kg" | "lbs";
+	weightStep: number;
+	exerciseName: string;
+	onSetPlanChange: (setPlanId: string, pattern: NonNullable<Pattern>) => void;
 };
 
-export const SetPlanSection: FC<Props> = ({ setPlans, weightUnit }) => {
+export const SetPlanSection: FC<Props> = ({
+	setPlans,
+	weightUnit,
+	weightStep,
+	exerciseName,
+	onSetPlanChange,
+}) => {
 	if (setPlans.length === 0) return null;
 	return (
 		<ol className="flex flex-col">
@@ -16,8 +25,13 @@ export const SetPlanSection: FC<Props> = ({ setPlans, weightUnit }) => {
 				<li key={setPlan.id}>
 					<SetPlanRow
 						index={index}
-						params={setPlan.params}
+						pattern={setPlan.pattern}
 						weightUnit={weightUnit}
+						weightStep={weightStep}
+						edit={{
+							title: `${exerciseName} ${index + 1}セット目`,
+							onChange: (pattern) => onSetPlanChange(setPlan.id, pattern),
+						}}
 					/>
 				</li>
 			))}
